Move `with_input.rs` into its own module
